refactored the main text area into a standalone class
This is in anticipation of beefing up its functionality e.g. with @-mentions
